docker_django/requirements.txt
Django==1.11.7
django-crispy-forms==1.7.0

psycopg2==2.7.3.2
